+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Home Control Arduino v9 (2023-12) ##
+
+* Curtain device with open and close pin added
+* milis() overflow correctly fixed
+
 ## Home Control Arduino v8 (2022-03) ##
 
 * Relay inverted option support